% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{gg_barplot_pct}
\alias{gg_barplot_pct}
\title{Barplot with labels}
\usage{
gg_barplot_pct(dataframe, xvar, fillvar, title = NULL)
}
\arguments{
\item{dataframe}{disaggregated data to plot}

\item{xvar}{variable to appear on the x axis; percents are computed based on this}

\item{fillvar}{variable to be used for color}

\item{title}{optional string to be set as the chart title}
}
\description{
Function creating a visualization based on \code{geom_bar()} with count and percent labels.
}
\examples{
data(diamonds, package = 'ggplot2')
gg_barplot_pct(diamonds, xvar = cut, fillvar = color, title = 'Diamond color and cut')
}
